printk: Do not disable preemption when calling consoles in RT
console_unlock() may call consoles that call spin_lock(), where spin_lock is
a mutex in PREEMPT_RT. Although mainline may be worried about preempted
holders of the console lock from preventing printk() output, the RT kernel
does not.

Use migrate_disable() to disable preemption when !RT, but we really don't
care if the task stays on the CPU or not for RT. It just has to be able to
schedule.

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>