thermal/int340x_thermal: Export OEM vendor variables
commit 006f006f1e5c48e5ddd6515eab2c9a7b27ce1144 upstream.

The platform vendor may expose an array of OEM-specific values to be used
in determining DPTF policy. These are obtained via the ODVP method, and
then simply exposed in sysfs. In addition, they are updated when a
notification is received or when the DPTF policy is changed by userland.

Conflicts:
	drivers/thermal/intel/int340x_thermal/int3400_thermal.c

Signed-off-by: Matthew Garrett <mjg59@google.com>
Tested-by: Pandruvada, Srinivas <srinivas.pandruvada@linux.intel.com>
Signed-off-by: Zhang Rui <rui.zhang@intel.com>
Link: https://lore.kernel.org/r/20200414020953.255364-2-matthewgarrett@google.com
Signed-off-by: Liwei Song <liwei.song@windriver.com>