thunderbolt: Tear down DP tunnels when suspending
commit 81a2e3e49f1f89cfe30d354b3a5fba27e2e59d91 upstream.

DP tunnels do not need the same kind of treatment as others because they
are created based on hot-plug events on DP adapter ports, and the
display stack does not need the tunnels to be enabled when resuming from
suspend. Also Tiger Lake Thunderbolt controller sends unplug event on D3
exit so this avoids that as well.

Signed-off-by: Mika Westerberg <mika.westerberg@linux.intel.com>
Signed-off-by: Liwei Song <liwei.song@windriver.com>