perf,mm: Handle non-page-table-aligned hugetlbfs
commit 51b646b2d9f84d6ff6300e3c1d09f2be4329a424 upstream.

A limited nunmber of architectures support hugetlbfs sizes that do not
align with the page-tables (ARM64, Power, Sparc64). Add support for
this to the generic perf_get_page_size() implementation, and also
allow an architecture to override this implementation.

This latter is only needed when it uses non-page-table aligned huge
pages in its kernel map.

Signed-off-by: Peter Zijlstra (Intel) <peterz@infradead.org>
Signed-off-by: Liwei Song <liwei.song@windriver.com>