thunderbolt: Add connection manager specific hooks for USB4 router operations
commit 9490f71167feba55349e33854f5e51a1a3af9e8c upstream.

Intel USB4 host routers that run the firmware based connection manager
(ICM) may implement a proxy for USB4 router operations. This is to avoid
the firmware to race with the OS driver, as both may need to run these
operations.

This adds two new connection manager specific callbacks which, if
provided, get called instead of the native USB4 router operation.

Signed-off-by: Mika Westerberg <mika.westerberg@linux.intel.com>
Signed-off-by: Liwei Song <liwei.song@windriver.com>