media: bcm2835-unicam: Return early from stop_streaming() if stopped
commit  52e76d9778922c7474b2b48ddad74cd9609959df from
https://github.com/raspberrypi/linux.git rpi-5.4.y

clk_disable_unprepare() is called unconditionally in stop_streaming().
This is incorrect in the cases where start_streaming() fails, and
unprepares all clocks as part of the failure cleanup. To avoid this,
ensure that clk_disable_unprepare() is only called in stop_streaming()
if the clocks are in a prepared state.

Signed-off-by: Naushir Patuck <naush@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>