s32gen1: qspi: Improve QSPI read performance
commit 862780563667294674804649acf9d90f41894482 from
https://source.codeaurora.org/external/autobsps32/linux

During QSPI read operation, using AHB buffer,
the call to 'memcpy_fromio' is not necessary.
Moreover, it leads to a performance decrease of up to 40%.

The solution is to use a simple 'memcpy' call, since
QSPI is memory mapped.

Also, we add performance indicators when performing a
QSPI read operation, measured in MiB/s.

Issue: ALB-6395
Signed-off-by: Ciprian Marian Costea <ciprianmarian.costea@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>