pcie:s32gen1: Remove SoC version check when enabling PCIe
commit a5e9d7ddb69bd917ef20e8460c6dd540a6162f1f from
https://source.codeaurora.org/external/autobsps32/linux

This check is only valid for rev 1.0.0 S32G274A SoCs, which had
PCIe issues.
This revision is now obsolete and only existed on some of the
early S32G274AEVBs (and which did not get outside NXP).
More, the check is still performed in u-boot, which is enough.

Issue: ALB-4529

Signed-off-by: Ionut Vicovan <Ionut.Vicovan@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>