mtd: spi-nor: change the performance info print to dev_dbg
In commit e952a9418b16 ("s32gen1: qspi: Improve QSPI read performance"),
it introduced a performance indicators when performing a QSPI read
operation, while it is no much help for the flash function and
may cause some flash test fail due to the tons of infos. So downgrade
dev_info to dev_dbg to fix it.

Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>