driver core: auxiliary bus: Enable by default
This commit enables CONFIG_AUXILIARY_BUS by default. This is necessary,
because StarlingX does not enable any kernel modules that 'select'
CONFIG_AUXILIARY_BUS, which causes the kernel's build system to
automatically disable this option.

However, StarlingX has out-of-tree users of this functionality (such as
the 'ice' and 'mlnx-ofa_kernel' out-of-tree kernel driver packages),
hence the need for this change. Prior to this commit, the aforementioned
out-of-tree kernel drivers would use their embedded/bundled copy of the
same functionality, which would cause kernel symbol conflicts at
run-time when users attempted to load both ice and mlnx-ofa_kernel
drivers.

Signed-off-by: M. Vefa Bicakci <vefa.bicakci@windriver.com>